# web/swagger/swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"RootController_root","parameters":[],"responses":{"200":{"description":""}},"tags":["Root"]}},"/users":{"get":{"description":"Get all users' profile","operationId":"UserController_getAll","parameters":[],"responses":{"200":{"description":"Successful","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}}}}},"security":[{"admin":[]}],"summary":"Get all profiles","tags":["User"]},"post":{"description":"Register a profile for existed user in Supabase's auth database. (User only)","operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"User profile created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"409":{"description":"Conflict","content":{"application/json":{"example":{"response":{"message":"User profile is already existed, cannot create a new one","error":"Conflict","statusCode":409},"status":409,"options":{},"message":"User profile is already existed, cannot create a new one","name":"ConflictException"}}}}},"security":[{"admin":[]},{"user":[]}],"summary":"Register profile","tags":["User"]}},"/users/{id}":{"get":{"description":"Get user profile by ID","operationId":"UserController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"403":{"description":"Not allowed","content":{"application/json":{"example":{"response":{"message":"You are only allowed to get your own profile","error":"Forbidden","statusCode":403},"status":403,"options":{},"message":"You are only allowed to get your own profile","name":"ForbiddenException"}}}},"404":{"description":"User not found","content":{"application/json":{"example":{"response":{"message":"User not found","error":"Not Found","statusCode":404},"status":404,"options":{},"message":"User not found","name":"NotFoundException"}}}}},"security":[{"user":[]},{"admin":[]}],"summary":"Get profile","tags":["User"]},"put":{"description":"Update user's profile, need to send all properties.","operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"204":{"description":"User profile updated successfully"},"403":{"description":"Not allowed","content":{"application/json":{"example":{"response":{"message":"You are not allowed to update other user's profile","error":"Forbidden","statusCode":403},"status":403,"options":{},"message":"You are not allowed to update other user's profile","name":"ForbiddenException"}}}},"404":{"description":"User doesn't exist to be updated","content":{"application/json":{"example":{"response":{"message":"User doesn't exist to be updated","error":"Not Found","statusCode":404},"status":404,"options":{},"message":"User doesn't exist to be updated","name":"NotFoundException"}}}}},"security":[{"user":[]},{"admin":[]}],"summary":"Update profile","tags":["User"]},"patch":{"description":"Update paritial user's profile. User ID is required for admin request","operationId":"UserController_updateParitial","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateParitialUserDto"}}}},"responses":{"200":{"description":"User profile updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"403":{"description":"Not allowed","content":{"application/json":{"example":{"response":{"message":"You are not allowed to paritially update other user's profile","error":"Forbidden","statusCode":403},"status":403,"options":{},"message":"You are not allowed to paritially update other user's profile","name":"ForbiddenException"}}}},"404":{"description":"User doesn't exist to be updated","content":{"application/json":{"example":{"response":{"message":"User doesn't exist to be updated","error":"Not Found","statusCode":404},"status":404,"options":{},"message":"User doesn't exist to be updated","name":"NotFoundException"}}}}},"security":[{"user":[]},{"admin":[]}],"summary":"Update paritial profile","tags":["User"]},"delete":{"description":"Delete user profile with user id","operationId":"UserController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"User deleted successfully"},"403":{"description":"Not allowed","content":{"application/json":{"example":{"response":{"message":"You are not allowed to delete other user's profile","error":"Forbidden","statusCode":403},"status":403,"options":{},"message":"You are not allowed to delete other user's profile","name":"ForbiddenException"}}}},"404":{"description":"User doesn't exist to be deleted","content":{"application/json":{"example":{"response":{"message":"User doesn't exist to be deleted","error":"Not Found","statusCode":404},"status":404,"options":{},"message":"User doesn't exist to be deleted","name":"NotFoundException"}}}}},"security":[{"user":[]},{"admin":[]}],"summary":"Delete profile","tags":["User"]}},"/users/{id}/sources":{"get":{"description":"Get all sources of user with user id","operationId":"UserController_getSources","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Successful","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SourceEntity"}}}}},"403":{"description":"Not allowed","content":{"application/json":{"example":{"response":{"message":"You are not allowed to get other user's sources","error":"Forbidden","statusCode":403},"status":403,"options":{},"message":"You are not allowed to get other user's sources","name":"ForbiddenException"}}}},"404":{"description":"User doesn't exist to have sources","content":{"application/json":{"example":{"response":{"message":"User doesn't exist to have sources","error":"Not Found","statusCode":404},"status":404,"options":{},"message":"User doesn't exist to have sources","name":"NotFoundException"}}}}},"security":[{"user":[]},{"admin":[]}],"summary":"Get sources of user","tags":["User"]}},"/users/{id}/tickets":{"get":{"description":"Get all tickets of user with user id","operationId":"UserController_getTickets","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Successful","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TicketEntity"}}}}},"403":{"description":"Not allowed","content":{"application/json":{"example":{"response":{"message":"You are not allowed to get other user's sources","error":"Forbidden","statusCode":403},"status":403,"options":{},"message":"You are not allowed to get other user's sources","name":"ForbiddenException"}}}},"404":{"description":"User doesn't exist to have sources to have tickets","content":{"application/json":{"example":{"response":{"message":"User doesn't exist to have sources to have tickets","error":"Not Found","statusCode":404},"status":404,"options":{},"message":"User doesn't exist to have sources to have tickets","name":"NotFoundException"}}}}},"security":[{"user":[]},{"admin":[]}],"summary":"Get tickets of user","tags":["User"]}},"/payment/tranfer-money":{"post":{"operationId":"PaymentController_tranferMoney","parameters":[],"responses":{"201":{"description":""}},"tags":["Payment"]}},"/payment/vnpay/create":{"post":{"description":"Generate payment link for VNPay","operationId":"VnpayController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTopUpDto"}}}},"responses":{"200":{"description":"VNPAY payment link","content":{"application/json":{"example":"https://sandbox.vnpayment.vn/paymentv2/vpcpay.html?vnp_Amount=1806000&vnp_Command=pay&vnp_CreateDate=20210801153333&vnp_CurrCode=VND&vnp_IpAddr=127.0.0.1&vnp_Locale=vn&vnp_OrderInfo=Thanh+toan+don+hang+%3A5&vnp_OrderType=other&vnp_ReturnUrl=https%3A%2F%2Fdomainmerchant.vn%2FReturnUrl&vnp_TmnCode=DEMOV210&vnp_TxnRef=5&vnp_Version=2.1.0&vnp_SecureHash=3e0d61a0c0534b2e36680b3f7277743e8784cc4e1d68fa7d276e79c23be7d6318d338b477910a27992f5057bb1582bd44bd82ae8009ffaf6d141219218625c42"}}}},"security":[{"user":[]}],"summary":"Generate payment link","tags":["Vnpay"]}},"/payment/vnpay/vnpay-ipn":{"get":{"description":"This is for VNPay to callback, not for the user / admin.","operationId":"VnpayController_ipnCallback","parameters":[],"responses":{"200":{"description":""}},"summary":"IPN URL for VNPay","tags":["Vnpay"]}},"/sources/{id}/tickets":{"get":{"description":"Get all tickets of source with source id","operationId":"SourceController_getTickets","parameters":[],"responses":{"200":{"description":"successful","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TicketEntity"}}}}},"404":{"description":"Source not found","content":{"application/json":{"example":{"response":{"message":"Source not found","error":"Not Found","statusCode":404},"status":404,"options":{},"message":"Source not found","name":"NotFoundException"}}}}},"security":[{"user":[]}],"summary":"Get all tickets of source","tags":["Source"]}},"/sources":{"post":{"description":"Create source for user with user id","operationId":"SourceController_createSource","parameters":[],"responses":{"200":{"description":"Successfull","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SourceEntity"}}}},"404":{"description":"User doesn't exist to be created source","content":{"application/json":{"example":{"response":{"message":"User doesn't exist to be created source","error":"Not Found","statusCode":404},"status":404,"options":{},"message":"User doesn't exist to be created source","name":"NotFoundException"}}}}},"security":[{"user":[]}],"summary":"Create source in user","tags":["Source"]}},"/plans":{"get":{"operationId":"PlanController_findAll","parameters":[{"name":"days","required":true,"in":"query","schema":{"type":"number"}},{"name":"isDisabled","required":true,"in":"query","schema":{"type":"boolean"}}],"responses":{"200":{"description":""}},"tags":["Plan"]}},"/plans/{id}":{"get":{"operationId":"PlanController_getPlanById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Plan"]}},"/tickets/{id}":{"get":{"operationId":"TicketController_getTicketById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Ticket"]}},"/tickets":{"post":{"operationId":"TicketController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTicketDto"}}}},"responses":{"201":{"description":""}},"tags":["Ticket"]}},"/methods/{id}":{"get":{"operationId":"MethodController_getMethodById","parameters":[],"responses":{"200":{"description":""}},"tags":["Method"]}},"/admins/create":{"post":{"operationId":"AdminController_createAdmin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAdminDto"}}}},"responses":{"201":{"description":"Successfully logged in"}},"tags":["Admin"]}},"/admins":{"get":{"operationId":"AdminController_getAdmins","parameters":[],"responses":{"200":{"description":"Get admins' information","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AdminEntity"}},"example":[{"username":"GuessMe","password":"StealMe!!"},{"username":"foo","password":"bar"}]}}}},"security":[{"admin":[]}],"tags":["Admin"]}},"/admins/login":{"post":{"operationId":"AdminController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginAdminDto"}}}},"responses":{"200":{"description":"Login the admin","content":{"application/json":{"example":{"accessToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJhZG1pbiIsImlhdCI6MTc0MjgzMDI3NywiZXhwIjoxNzQyODMzODc3fQ.Uqnp518LNF6ZRVmrIy97c165XPAo5-s44UV0cTlS6f4","refreshToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJhZG1pbiIsImlhdCI6MTc0MjgzMDI3NywiZXhwIjoxNzQzNDM1MDc3fQ.KdWa6w76vn1GTz0sPM8bCzmoneJxEBsRMtdP5WlfdnE"}}}},"401":{"description":"Who the fuck are you?"}},"security":[{"admin":[]}],"tags":["Admin"]}},"/admins/refresh":{"post":{"operationId":"AdminController_refreshToken","parameters":[],"responses":{"200":{"description":"Successfully refreshed the token","content":{"application/json":{"example":{"accessToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJhZG1pbiIsImlhdCI6MTc0MjgzMDI3NywiZXhwIjoxNzQyODMzODc3fQ.Uqnp518LNF6ZRVmrIy97c165XPAo5-s44UV0cTlS6f4","refreshToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJhZG1pbiIsImlhdCI6MTc0MjgzMDI3NywiZXhwIjoxNzQzNDM1MDc3fQ.KdWa6w76vn1GTz0sPM8bCzmoneJxEBsRMtdP5WlfdnE"}}}},"401":{"description":"Who the fuck are you?"}},"security":[{"admin":[]}],"tags":["Admin"]}}},"info":{"title":"CoinHub","description":"CoinHub API docs","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"user":{"scheme":"bearer","bearerFormat":"JWT","type":"http"},"admin":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"UserEntity":{"type":"object","properties":{"id":{"type":"string"},"createdAt":{"type":"string"},"fullname":{"type":"string"},"birthDate":{"type":"string"},"citizenId":{"type":"string"},"avatar":{"type":"string","nullable":true},"address":{"type":"string","nullable":true}},"required":["id","createdAt","fullname","birthDate","citizenId","avatar","address"]},"CreateUserDto":{"type":"object","properties":{"fullname":{"type":"string"},"birthDate":{"type":"string","example":"3/31/2001"},"citizenId":{"type":"string","description":"Citizen ID with 12 digits","example":"077002455001"},"avatar":{"type":"string","description":"Avatar URL","example":"https://avatars.githubusercontent.com/u/86353526?v=4"},"address":{"type":"string"},"phoneNumber":{"type":"string","description":"Phone number with 10 digits","example":"0945678910"}},"required":["fullname","birthDate","citizenId"]},"UpdateUserDto":{"type":"object","properties":{"fullname":{"type":"string"},"birthDate":{"type":"string","example":"3/31/2001"},"citizenId":{"type":"string","description":"Citizen ID with 12 digits","example":"077002455001"},"avatar":{"type":"string","description":"Avatar URL","example":"https://avatars.githubusercontent.com/u/86353526?v=4"},"address":{"type":"string"},"phoneNumber":{"type":"string","description":"Phone number with 10 digits","example":"0945678910"}},"required":["fullname","birthDate","citizenId"]},"UpdateParitialUserDto":{"type":"object","properties":{"fullname":{"type":"string"},"birthDate":{"type":"string","example":"3/31/2001"},"citizenId":{"type":"string","description":"Citizen ID with 12 digits","example":"077002455001"},"avatar":{"type":"string","description":"Avatar URL","example":"https://avatars.githubusercontent.com/u/86353526?v=4"},"address":{"type":"string"},"phoneNumber":{"type":"string","description":"Phone number with 10 digits","example":"0945678910"}}},"SourceEntity":{"type":"object","properties":{"id":{"type":"string"},"balance":{"type":"string"}},"required":["id","balance"]},"TicketEntity":{"type":"object","properties":{"id":{"type":"string"},"openedDate":{"format":"date-time","type":"string"},"closedDate":{"type":"object"}},"required":["id","openedDate","closedDate"]},"CreateTopUpDto":{"type":"object","properties":{"type":{"type":"string","enum":["VNPAY","MOMO","ZALOPAY"]},"returnUrl":{"type":"string","description":"Return URL to client app","example":"coinhub://home"},"amount":{"type":"number","description":"The amount of money","example":100000},"sourceDestination":{"type":"string"},"ipAddress":{"type":"string"}},"required":["type","returnUrl","amount","sourceDestination","ipAddress"]},"CreateTicketDto":{"type":"object","properties":{"sourceId":{"type":"string"},"methodId":{"type":"string"}},"required":["sourceId","methodId"]},"CreateAdminDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"AdminEntity":{"type":"object","properties":{"username":{"type":"string"}},"required":["username"]},"LoginAdminDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]}}}}